Pass parameters to LogTailer in childspec
Elixir 1.6 reported that it was creating a default `init/1` function for
the LogTailer.* modules. That function isn't used since the
`LogTailer.start_link` call passes its module as the first parameter.
One fix would be to `defdelegate` the `init` function to suppress the
warning. But this is a little confusing since it would never be called
at runtime.

To avoid any confusion on which callbacks get called in which module, I
just parameterized the child spec. This also fixes the Elixir 1.6 `init`
warning.