Fix dialyzer error with port binary name
The port's executable was specified as an Elixir string instead of an
Erlang string. This works, but Dialyzer complains. This change fixes it.
The fix looks a little uglier since it hardcodes the path separator, but
I justified the change since it's how the uevent port executable is
referenced (consistency).