Run uevent discovery asynchronously
Previously the discovery was being done synchronously in the UEvent
GenServer. This would cause uevents messages from the port process to
queue up in the mailbox. The once discovery was done, they'd be
processed in a burst. With this change, the discovery happens in another
process. UEvents get processed as they come in. This opens up some
parallelism. The most obvious difference is that kernel modules get
loaded more quickly.