Find the uevent files and then add
This changes the uevent discovery process to find all of the uevent
files and then to write "add" to them. Previously, "add" was written to
them while finding them.

The reason that this is ok is because when the discovery process is
started, the uevent port has been started. Therefore, if a device were
to get added during the file discovery process, it would already have
sent a uevent that the uevent port would have seen.

This change appears to noticeably speed up discovery time when combined
with #107 on the Raspberry Pi Zero.