Ignore SYNTH_UUID in uevent reports
SYNTH_UUID is a relatively new addition to the uevent reports that lets
applications distinguish between real events and those events that were
generated by writing the uevent file. The distiction between synthetic
and real events isn't used and this information isn't useful in
SystemRegistry, so strip it off. This seems to reduce SystemRegistry
reductions by 10%.