Don't exit when an overload occurs
I can only reproduce an overload on a 32-core x86_64 machine. What
happens is that libmnl returns an error from the recv call that
indicates that the kernel internally has dropped a notification. The
overload is triggered in the initial scan of hardware. It seems like it
should be impossible to trigger in any other real hardware insertion and
removal scenario.

Before this fix, the behavior was to exit the port process. This would
cause it to be re-started due to supervision. Since the hardware scan
must happen again since devices could have been inserted between the
crash and the restart, the scan is kicked off and the port process exits
again. Restart intensity is reached and total failure. Losing one or two
notifications bothers me greatly, but it seems like a far better
solution than crashing.

I suspect that this has not been seen in the field since the number of
devices on most embedded systems is far less than on this 32-core
machine.