Move device discovery to uevent port executable
There was a race condition (that I never saw) where if uevent discovery
happened before the uevent port executable started that discovery
messages could be lost. This moves the discovery code to after libmnl
has been initialized so that the socket will always be open to receive
events. Since discovery always has to be done when the uevent listener
is started to avoid missing any added devices, this hardcodes that call
in C.

This is a potentially backwards incompatible change if anyone was
manually forcing device discovery events using
`Nerves.Runtime.Device`. That should never have been necessary, so if
this was done, there was almost certainly another issue and perhaps it
was due to the race condition that this fixes.