Load modules from /etc/modules
Auto-load modules listed in the `/etc/modules` file. This makes it
possible for systems to specify (and users to override) the load of
kernel modules that may not be detectable via `uevent` messages.

Previously these modules were compiled into the kernel, but this does
not support cases where they could be optional. For example, it would
be nice to have a few USB gadget driver options without hardcoding them
into the kernel. This lets systems be made where the default case loads
a Ethernet/serial gadget driver (for example), but that can be disabled
by overriding `/etc/modules` with a different set of drivers.