Clarify docs for using Nerves.Runtime.KV.Mock
I noticed that the old way of setting mock data is not longer supported. So this changes the docs with the correct way of specifying KV.Mock data.

Also adds `@callback put(state)` module attribute in `Nerves.Runtime.KV` since that is now a function which will be called on the dynamically defined module and should be expected in the contract.