Add /dev/kmsg parsing; refactor syslog parser
This adds a parser for the /dev/kmsg format. since there's a lot of
reused code between kmsg and syslog parsing, this also refactors the
syslog parser to support the sharing. Many tests were added to exercise
some bugs that were found on inspection and to better cover the code.