Support uboot_env 0.2.0
uboot_env 0.2.0 contains a breaking change from 0.1.1, but the breaking
change does not affect nerves_runtime. This commit makes it possible to
use either version.

The main reason to update to 0.2.0 is that it significantly reduces the
amount of memory garbage that gets generated when parsing and writing
environment blocks. On one production system the garbage created was in
the low megabytes which one easily bump any process that called
uboot_env to the top of the list.