Autovalidate fw on startup
Currently only a few systems used `nerves_fw_validated` and its  a
manual process in `fwup.conf` to setup and check in the context of
autoreverts. Then that variable gets checked on next boot to determine
if a revert should happen.

However, I think this is a little too early in the startup (i.e. we
immediately set this variable, but something later in the startup but
before runtime fails).

So this sets it during initialization of `Nerves.Runtime` to help catch
that zone of errors that might otherwise be missed. If we're to this
initialization, we're pretty confident the fw is good.

It also allows disabling autovalidation in the config so that the
application can handle the logic check even later.