Revert "Autovalidate fw on startup"
This reverts commit f2ba03b34957ea0cec72a734e443df7ecfd62cfe.

See #180 and #184 for the discussion. The decision was made that users
should call `Nerves.Runtime.validate_firmware/0` in their applications.
This is a one-liner. Currently this is NOT required in any of the
official systems, but the plan is to make it required (perhaps by adding
to to new project generator) so that auto-failback on bad firmware can
be a feature that more users can have out-of-the-box.