Support U-Boot bootcount/upgrade_available
U-Boot supports running alternative boot commands if the main one fails
too many times. This can be used for trying out firmware updates and has
a couple advantages: 1. It's easier to talk to people who use U-Boot,
but not Nerves, 2. You can set a new firmware to be tried more than once
before reverting, and 3. The U-Boot script is simpler. The second reason
is interesting if aspects of validating a firmware image are out of your
control and it's better to try it a few times "just in case".

To use this feature, you need to update your U-Boot scripts and make
sure that the bootcount feature is enabled. The fwup script also needs
to be adjusted to set "upgrade_available" and "bootcount".

This change has no effect if you're not using the U-Boot "bootcount"
feature.