Change firmware_valid? to be valid when not used
If you think real hard when the firmware validity feature isn't used,
the firmware is implicitly valid. I.e., a reboot keeps the firmware
rather than reverting.

This changes the function to return `true` when it's not valid.

The immediate effect of this is that Raspberry Pi's which don't
auto-revert back on bad firmware will report `true` now.