Set shell_history on reformat
After a fresh SD card burn, the application partition may not be
formatted which means the shell history will fail to start preventing
history being saved until a reboot.

This change detects if the application partition needed to be formatted
and resets the shell history env var to enable history again once
the disk was successfully formatted and mounted. This will allow
history to be saved even on first boot.