Remove all references to SystemRegistry
We've advocated not using SystemRegistry for years now and it has been
disabled by default for a long time. This completely removes it so that
we can drop maintaining it.

If you're seeing this commit and are using the SystemRegistry
integration to get insertion and removal ports, an alternative is after
this commit.