Move modprobe invocation to C
This saves a trip from C -> Elixir and back to run modprobe for
dynamically loading kernel modules. Since it's unknown whether modprobe
should be called for any new device, it always has to be called. This
ends up being 50-100 invocations or more depending on the device.

GRiSP 2 has a relatively small number of devices (~50), but moving this
to C results in about 1 second reduction of boot time. Boot time here is
for when the critical applications are all started (Nerves.Runtime,
VintageNet, etc.)

Calling in C has a number of advantages as well. First of all, stdio can
be redirected to `/dev/null` since we know that we're going to throw
away any output anyway. Elixir must capture this output and then toss
it. Secondly, starting processes is cheaper in C anyway, since Erlang
has to do some tricks to avoid forks copying a lot of process memory.