Update device scan for determinism and container first
This is mostly a cleanup of the ordering of events in the system, but it
might fix a subtle issue.

The general idea behind this code is to trigger uevent reports to be
sent over netlink. We may have started listening for reports after
events were sent, so to be safe, everything needs to be told to send an
event again. The way this is done is to scan `/sys/devices` and write
`add` to every `uevent` file.

The original file processing order matched whatever was on the
filesystem. This does not look deterministic (and that makes sense that
it wouldn't since that's extra work). Files were also processed in post
order, so USB devices would get their `add`'s written to them before the
USB bus devices that they're on, etc.

This commit changes the logic in two ways:

1. Parent `uevent` files are processed before child `uevent` files. This
   seems intuitively what users would expect.
2. Directories are scanned in alphabetical order. This is for
   determinism. It "feels" good for the scan the be deterministic if
   future oddities in this code are found.

This fixes a subtle kernel module load order issue: Since Linux kernel
modules are loaded based on uevents, the order they're delivered can
matter. If a driver depends on another driver to be loaded to trigger
off the "modalias", sending the child uevent first appears to not work.
This can be reproduced with a RT5370 WiFi module. This "worked" prior to
the `modprobe` optimization in c4b055c. The order looks like it should
have a problem, but maybe there's a timing element. At any rate, this
new ordering seems to make a whole lot more sense (to me) and drivers
reliably load.