Batch up modprobe invocations
Each call to modprobe is a fork and exec. Then after that, modprobe
needs to read the modalias database. This change batches up calls to
modprobe so that they can be done in one go.

Here are some test runs:

BBB clone 1 -> 78 modprobes became 9
BBB clone 2 -> 83 modprobes became 11
GRiSP2      -> 73 modprobes became 8
RPi0        -> 51 modprobes became 5

This has a small effect on reducing boot time. Rough timing seemed to
make it less than a second.

Why is modprobe called so much? It's because there's no wait to know
whether a kernel module should be loaded when a device gets reported.
The only way to tell is to check the modalias against the kernel module
database.