Decouple and serialize reboot and poweroff calls
This moves the reboot and poweroff handling functions to a dedicated
GenServer to address a rare, but observed case where the supervision
tree for the GenServer that called `Nerves.Runtime.reboot` went down.
This ended up killing the reboot process midway and the device didn't
reboot.

The reboot/poweroff code is now in a singleton GenServer. This also
fixes the case where reboot gets called by multiple processes and there
are dualing reboots happening.