Update .circleci/config.yml
This brings in a few CI updates to hopefully make it easier to adjust and
use in the future. This matches what is currently being used in `property_tables`

Adding a new elixir/OTP version is now a simple tag addition to the matrix.
To mark a new tag as the latest, you only need to adjust the `&latest` anchor