Remove unnecessary mix target config
The Mix target that was used to build Nerves.Runtime was being capture
in the application configuration and turned into a string. This is
unncessary since it can be captured at compile-time. Additionally, it's
easier to use the Mix target if it's kept an atom.

Rather than try to maintain the past, I deleted
`Nerves.Runtime.target/0` so users will need to update their code if
they used it. However, since you had to go away and configure
`:nerves_runtime` manually in your code to use it, I suspect that it was
either never used or just luck that it worked.

The new function is `Nerves.Runtime.mix_target/0`.

This also does the clean to remove any hint that the Mix target is
stored in the application config.