Add kv_backend option for supplying alternate KV
The KV store used to be configurable, but it was very much a unit test
feature in how it was configured: add the `:modules` key with
`Nerves.Runtime.KV.Mock` options.

This change refactors this support by adding the concept of a KV
backend and renaming the "Mock" implementation to `InMemory` to signify
that it's an in-memory KV store.

While this introduces a new way of configuring the backend, the old way
still works so it's backwards compatible.

Since configuring the KV backend happens early and crashes are
unforgiving, this code handles misconfigurations by reverting to the
in-memory KV store.