Simplify and further bulletproof poweroff error cases
This cleans up the poweroff code that got a little more complicated with
the guarded poweroff feature. There's no GenServer running and the code
to prevent multiple calls to power off simultanously has been removed
since that doesn't seem to be a big risk. The poweroff is still run in
an process that's not linked to the caller to avoid supervision tree
restarts from halting a poweroff.

This also fixes a bug where control was returned to the caller after an
`:init.stop/0`. The new way is to wait 10 minutes and then call
`:erlang.halt/0` as a last resort. The guarded reboot/poweroff should
kick in before this, though.