Support ungraceful reboot and use for factory reset
Since the factory reset mechanism runs in the application environment
(rather than in the bootloader or some more pristine location), it has
to deal with other code accidentally undoing what it does. Graceful
reboot could consistently do this on some configurations and the result
was a really bad partial factory reset. This updates the factory reset
to invoke an immediate reboot feature that was added to Nerves Heart.