Remount ext4 app partition read-only on error
The current default behavior is to return filesystem corruption errors
and stay "writable". This seems reasonable, but there's been a history
of remounting read-only on error, so it was surprising when the failure
changed. This restores the failure mode. F2FS does not have a similar
option, so only set it for ext4.