Hide Dialyzer warnings when using mix_target/0
Dialyzer correctly raises unused code warnings on code paths that don't
match the current target (usually `:host`). The best fix is to move this
code to compile time to avoid dead code in projects.

The Dialyzer warnings aren't configurable to provide any help in this
area, and it makes this issue way more frustrating for amount of dead
code that would likely be removed. Therefore, just launder the setting
through `Function.identity/1` and Dialyzer stopps complaining.