Add a simple firmware auto-validator
Since it's going to be possible in the official Nerves systems to try
out firmware that needs logic to validate it, there needs to be a simple
way for new users. This is a really basic validator that validates a <10
seconds after the OTP release init script completes successfully.
Failure will revert after 15 minutes. I don't think there's a good
timeout for giving up, so this seems like more of an example that some
way of giving up is needed.