Add a simple firmware auto-validator
Since it's going to be possible in the official Nerves systems to try
out firmware that needs logic to validate it, there needs to be a simple
way for new users. This is a really basic validator that waits for all
OTP applications in the start script to start.

Applications not starting result in a reboot after 15 minutes which will
either revert or go through the process again. A warning message is
printed every minute to hopefully clue people into what's happening
since it's guaranteed that 15 minutes won't work for everyone.