Lazy load KV values
This defers the actual loading of the KV store out of the GenServer init
callback. This removes a possible place that could crash from
application init and defers an operation that will take some I/O time.

This also makes the forced refresh that happens after a generic put only
take place if `nerves_fw_active` was modified. Refreshing all the time
was certainly safe, but I'm not sure if an instance exists in practice
where it really was needed. Scoping it to the one important key seemed
like a safe change. It also lazily updates, since it you're doing this,
you really need to reboot asap before things get really confusing.